+++ plugin/src/main/groovy/com/novoda/staticanalysis/internal/detekt/DetektConfigurator.groovy
@@ -3,6 +3,7 @@ package com.novoda.staticanalysis.internal.detekt
 import com.novoda.staticanalysis.StaticAnalysisExtension
 import com.novoda.staticanalysis.Violations
 import com.novoda.staticanalysis.internal.Configurator
+import com.novoda.staticanalysis.internal.checkstyle.CollectCheckstyleViolationsTask
 import org.gradle.api.GradleException
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.Project
@@ -11,14 +12,16 @@ import org.gradle.api.Task
 class DetektConfigurator implements Configurator {
 
     private static final String DETEKT_PLUGIN = 'io.gitlab.arturbosch.detekt'
-    private static final String LAST_COMPATIBLE_DETEKT_VERSION = '1.0.0.RC8'
+    private static final String LAST_COMPATIBLE_DETEKT_VERSION = '1.0.0-RC10'
     private static final String DETEKT_NOT_APPLIED = 'The Detekt plugin is configured but not applied. Please apply the plugin in your build script.\nFor more information see https://github.com/arturbosch/detekt.'
     private static final String OUTPUT_NOT_DEFINED = 'Output not defined! To analyze the results, `output` needs to be defined in Detekt profile.'
     private static final String DETEKT_CONFIGURATION_ERROR = "A problem occurred while configuring Detekt. Please make sure to use a compatible version (All versions up to $LAST_COMPATIBLE_DETEKT_VERSION)"
+    private static final String XML_REPORT_NOT_ENABLED = 'XML report must be enabled. Please make sure to enable "reports.xml" in your Detekt configuration'
 
     private final Project project
     private final Violations violations
     private final Task evaluateViolations
+
 
     static DetektConfigurator create(Project project,
                                      NamedDomainObjectContainer<Violations> violationsContainer,
@@ -45,24 +48,51 @@ class DetektConfigurator implements Configurator {
             }
 
             def detekt = project.extensions.findByName('detekt')
+            setDefaultXmlReport(detekt)
             config.delegate = detekt
             config()
-            configureToolTask(detekt)
+
+            def collectViolations = configureToolTask(detekt)
+            evaluateViolations.dependsOn collectViolations
         }
     }
 
-    private void configureToolTask(detekt) {
-        def detektTask = project.tasks['detektCheck']
-        detektTask.group = 'verification'
+    private void setDefaultXmlReport(detekt) {
+        if (detekt.hasProperty('reports')) {
+            detekt.reports {
+                xml.enabled = true
+                xml.destination = new File(project.buildDir, 'reports/detekt/detekt.xml')
+            }
+        }
+    }
 
-        // evaluate violations after detekt
+    private CollectCheckstyleViolationsTask configureToolTask(detekt) {
+        def detektTask = project.tasks.findByName('detekt')
+        if (detektTask?.hasProperty('reports')) {
+            def reports = detektTask.reports
+            if (!reports.xml.enabled) {
+                throw new IllegalStateException(XML_REPORT_NOT_ENABLED)
+            }
+            return createCollectViolationsTask(
+                    violations,
+                    detektTask,
+                    reports.xml.destination,
+                    reports.html.destination
+            )
+        }
+
+        // Fallback to old Detekt versions
         def output = resolveOutput(detekt)
         if (!output) {
             throw new IllegalArgumentException(OUTPUT_NOT_DEFINED)
         }
-        def collectViolations = createCollectViolationsTask(violations, project.file(output))
-        evaluateViolations.dependsOn collectViolations
-        collectViolations.dependsOn detektTask
+        def outputFolder = project.file(output)
+        return createCollectViolationsTask(
+                violations,
+                project.tasks['detektCheck'],
+                new File(outputFolder, 'detekt-checkstyle.xml'),
+                new File(outputFolder, 'detekt-report.html')
+        )
     }
 
     private static resolveOutput(detekt) {
@@ -75,11 +105,13 @@ class DetektConfigurator implements Configurator {
         }
     }
 
-    private CollectDetektViolationsTask createCollectViolationsTask(Violations violations, File outputFolder) {
-        project.tasks.create('collectDetektViolations', CollectDetektViolationsTask) { task ->
-            task.xmlReportFile = new File(outputFolder, 'detekt-checkstyle.xml')
-            task.htmlReportFile = new File(outputFolder, 'detekt-report.html')
+    private CollectCheckstyleViolationsTask createCollectViolationsTask(Violations violations, detektTask, File xmlReportFile, File htmlReportFile) {
+        project.tasks.create('collectDetektViolations', CollectCheckstyleViolationsTask) { task ->
+            task.xmlReportFile = xmlReportFile
+            task.htmlReportFile = htmlReportFile
             task.violations = violations
+
+            task.dependsOn(detektTask)
         }
     }
 
@@ -87,8 +119,9 @@ class DetektConfigurator implements Configurator {
         final boolean isKotlin = project.plugins.hasPlugin('kotlin')
         final boolean isKotlinAndroid = project.plugins.hasPlugin('kotlin-android')
         final boolean isKotlinPlatformCommon = project.plugins.hasPlugin('kotlin-platform-common')
+        final boolean isKotlinMultiplatform = project.plugins.hasPlugin('org.jetbrains.kotlin.multiplatform')
         final boolean isKotlinPlatformJvm = project.plugins.hasPlugin('kotlin-platform-jvm')
         final boolean isKotlinPlatformJs = project.plugins.hasPlugin('kotlin-platform-js')
-        return isKotlin || isKotlinAndroid || isKotlinPlatformCommon || isKotlinPlatformJvm || isKotlinPlatformJs
+        return isKotlin || isKotlinAndroid || isKotlinPlatformCommon || isKotlinMultiplatform || isKotlinPlatformJvm || isKotlinPlatformJs
     }
 }