+++ plugin/src/test/groovy/com/novoda/staticanalysis/internal/detekt/DetektNewIntegrationTest.groovy
@@ -1,5 +1,6 @@
 package com.novoda.staticanalysis.internal.detekt
 
+
 import com.novoda.test.Fixtures
 import com.novoda.test.TestProject
 import com.novoda.test.TestProjectRule
@@ -11,18 +12,19 @@ import org.junit.runners.Parameterized
 import static com.novoda.test.LogsSubject.assertThat
 
 @RunWith(Parameterized.class)
-class DetektIntegrationTest {
+class DetektNewIntegrationTest {
 
     private static final String DETEKT_NOT_APPLIED = 'The Detekt plugin is configured but not applied. Please apply the plugin in your build script.'
-    private static final String OUTPUT_NOT_DEFINED = 'Output not defined! To analyze the results, `output` needs to be defined in Detekt profile.'
+    private static final String XML_REPORT_NOT_ENABLED = 'XML report must be enabled. Please make sure to enable "reports.xml" in your Detekt configuration'
+
 
     @Parameterized.Parameters(name = "{0} with Detekt: {1}")
     static Iterable rules() {
         return [
-                [TestProjectRule.forKotlinProject(), "1.0.0.RC6-2"],
-                [TestProjectRule.forAndroidKotlinProject(), "1.0.0.RC6-2"],
-                [TestProjectRule.forKotlinProject(), "1.0.0.RC8"],
-                [TestProjectRule.forAndroidKotlinProject(), "1.0.0.RC8"],
+                [TestProjectRule.forKotlinProject(), "1.0.0.RC9.2"],
+                [TestProjectRule.forAndroidKotlinProject(), "1.0.0.RC9.2"],
+                [TestProjectRule.forKotlinProject(), "1.0.0-RC10"],
+                [TestProjectRule.forAndroidKotlinProject(), "1.0.0-RC10"],
         ]*.toArray()
     }
 
@@ -30,20 +32,9 @@ class DetektIntegrationTest {
     public final TestProjectRule projectRule
     private final String detektVersion
 
-    DetektIntegrationTest(TestProjectRule projectRule, String detektVersion) {
+    DetektNewIntegrationTest(TestProjectRule projectRule, String detektVersion) {
         this.projectRule = projectRule
         this.detektVersion = detektVersion
-    }
-
-    @Test
-    void shouldFailBuildOnConfigurationWhenNoOutputNotDefined() {
-        def emptyConfiguration = detektWith(detektVersion, "")
-
-        def result = createProjectWithZeroThreshold(Fixtures.Detekt.SOURCES_WITH_WARNINGS)
-                .withToolsConfig(emptyConfiguration)
-                .buildAndFail('check')
-
-        assertThat(result.logs).contains(OUTPUT_NOT_DEFINED)
     }
 
     @Test
@@ -56,13 +47,27 @@ class DetektIntegrationTest {
     }
 
     @Test
+    void shouldFailBuildOnConfigurationWhenDetektConfiguredWithoutXmlReport() {
+        def result = projectRule.newProject()
+                .withPlugin("io.gitlab.arturbosch.detekt", detektVersion)
+                .withToolsConfig('''detekt {      
+                    reports {
+                        xml.enabled = false
+                    }
+                }''')
+                .buildAndFail('check')
+
+        assertThat(result.logs).contains(XML_REPORT_NOT_ENABLED)
+    }
+
+    @Test
     void shouldFailBuildWhenDetektWarningsOverTheThreshold() {
         def result = createProjectWithZeroThreshold(Fixtures.Detekt.SOURCES_WITH_WARNINGS)
                 .buildAndFail('check')
 
         assertThat(result.logs).containsLimitExceeded(0, 1)
         assertThat(result.logs).containsDetektViolations(0, 1,
-                result.buildFileUrl('reports/detekt-report.html'))
+                result.buildFileUrl('reports/detekt/detekt.html'))
     }
 
     @Test
@@ -72,7 +77,7 @@ class DetektIntegrationTest {
 
         assertThat(result.logs).containsLimitExceeded(1, 0)
         assertThat(result.logs).containsDetektViolations(1, 0,
-                result.buildFileUrl('reports/detekt-report.html'))
+                result.buildFileUrl('reports/detekt/detekt.html'))
     }
 
     @Test
@@ -89,7 +94,7 @@ class DetektIntegrationTest {
                 .build('check')
 
         assertThat(result.logs).containsDetektViolations(0, 1,
-                result.buildFileUrl('reports/detekt-report.html'))
+                result.buildFileUrl('reports/detekt/detekt.html'))
     }
 
     @Test
@@ -98,7 +103,7 @@ class DetektIntegrationTest {
                 .build('check')
 
         assertThat(result.logs).containsDetektViolations(1, 0,
-                result.buildFileUrl('reports/detekt-report.html'))
+                result.buildFileUrl('reports/detekt/detekt.html'))
     }
 
     @Test
@@ -145,29 +150,30 @@ class DetektIntegrationTest {
 
     private static String detektConfiguration(File input, String detektVersion) {
         detektWith(detektVersion, """
-            config = '${Fixtures.Detekt.RULES}' 
-            output = "\$buildDir/reports"
+            config = files('${Fixtures.Detekt.RULES}') 
             // The input just needs to be configured for the tests. 
             // Probably detekt doesn't pick up the changed source sets. 
             // In a example project it was not needed.
-            input = "${input}"
+            input = files("${input}")
         """)
     }
 
     private static String detektConfigurationWithoutInput(String detektVersion) {
         detektWith(detektVersion, """
-            config = '${Fixtures.Detekt.RULES}' 
-            output = "\$buildDir/reports"
+            config = files('${Fixtures.Detekt.RULES}') 
         """)
     }
 
-    private static String detektWith(String detektVersion, String mainProfile) {
+    private static String detektWith(String detektVersion, String configuration) {
         """
         detekt {      
-            version '${detektVersion}'
+            toolVersion '${detektVersion}'
             
-            profile('main') { 
-                ${mainProfile.stripIndent()}
+            ${configuration.stripIndent()}  
+
+            reports {
+                xml.enabled = true
+                html.enabled = true
             }
         }
         """